+++ pkg/resource/stack/deployment_test.go
@@ -84,6 +84,7 @@ func TestDeploymentSerialization(t *testing.T) {
 		nil,
 		nil,
 		nil,
+		"",
 	)
 
 	dep, err := SerializeResource(res, config.NopEncrypter)